+++ platform/linux-generic/Makefile.am
@@ -126,6 +126,7 @@ noinst_HEADERS = \
 		  include/odp_packet_dpdk.h \
 		  include/odp_packet_internal.h \
 		  include/odp_packet_io_internal.h \
+		  include/odp_parse_internal.h \
 		  include/odp_socket_common.h \
 		  include/odp_packet_io_stats_common.h \
 		  include/odp_packet_io_stats.h \
@@ -202,6 +203,7 @@ __LIB__libodp_linux_la_SOURCES = \
 			   odp_packet_vector.c \
 			   odp_packet_flags.c \
 			   odp_packet_io.c \
+			   odp_parse.c \
 			   odp_pkt_queue.c \
 			   odp_pool.c \
 			   odp_queue_basic.c \
