+++ platform/linux-generic/pktio/loop.c
@@ -18,6 +18,7 @@
 #include <odp/api/plat/packet_flag_inlines.h>
 #include <odp/api/plat/queue_inlines.h>
 
+#include <odp_parse_internal.h>
 #include <odp_classification_internal.h>
 #include <odp_debug_internal.h>
 #include <odp_errno_define.h>
