+++ platform/linux-generic/pktio/socket.c
@@ -15,6 +15,7 @@
 #include <odp/api/ticketlock.h>
 
 #include <odp_socket_common.h>
+#include <odp_parse_internal.h>
 #include <odp_packet_internal.h>
 #include <odp_packet_io_internal.h>
 #include <odp_packet_io_stats.h>
