+++ platform/linux-generic/pktio/socket_mmap.c
@@ -16,6 +16,7 @@
 #include <odp/api/plat/packet_inlines.h>
 
 #include <odp_socket_common.h>
+#include <odp_parse_internal.h>
 #include <odp_packet_internal.h>
 #include <odp_packet_io_internal.h>
 #include <odp_packet_io_stats.h>
